pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.setuptools.package-data]
openseries = ["*.json", "*.png"]

[project]
name = "openseries"
version = "0.10.2"
description = "Package for simple financial time series analysis."
readme = "README.md"
license = { text = "BSD 3-Clause License" }
authors = [
    { name = "Martin Karrin", email = "martin.karrin@captor.se" }
]
requires-python = ">=3.10"
dependencies = [
    "holidays>=0.18",
    "jsonschema>=4.17.3",
    "numpy>=1.24.1",
    "pandas>=1.5.2",
    "plotly>=5.11.0",
    "python-dateutil>=2.8.2",
    "python-stdnum>=1.18",
    "requests>=2.28.1",
    "scipy>=1.9.3",
    "statsmodels>=0.13.5"
]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "Natural Language :: English",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
]
keywords = [
    "python",
    "python3",
    "plotly",
    "pandas",
    "finance",
    "fintech",
    "data-science",
    "timeseries",
    "timeseries-data",
    "timeseries-analysis",
    "investment",
    "investment-analysis",
    "investing"
]

[project.optional-dependencies]
test = [
    "coverage",
    "pytest",
    "pytest-split"
]
dev = [
    "coverage",
    "pytest",
    "black==23.1.0",
    "flake8==6.0.0",
    "flake8-black"
]
build = [
    "build",
    "coverage",
    "pytest",
    "twine",
    "wheel"
]

[project.urls]
repository = "https://github.com/CaptorAB/OpenSeries"

[tool.black]
line-length = 88